requirements.txt
Flask
gunicorn
mysql-connector-python